backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.10.0
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pypdf==5.1.0
python-docx==1.1.2
pytest==8.3.3
